+++ src/planets/Mars.ts
@@ -5,6 +5,7 @@ import {observationCalc} from '../utils';
 import {DIAMETER_MARS} from '../constants/diameters';
 import IEclipticSphericalCoordinates from '../coordinates/interfaces/IEclipticSphericalCoordinates';
 import {normalizeAngle} from '../utils/angleCalc';
+import {getApparentMagnitudeMars} from '../utils/magnitudeCalc';
 
 export default class Mars extends Planet {
     public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates> {
@@ -35,5 +36,13 @@ export default class Mars extends Planet {
         const distance = await this.getApparentDistanceToEarth();
 
         return observationCalc.getAngularDiameter(distance, DIAMETER_MARS);
+    }
+
+    public async getApparentMagnitude(): Promise<number> {
+        const coordsHelio = await this.getHeliocentricEclipticSphericalDateCoordinates();
+        const coordsGeo = await this.getGeocentricEclipticSphericalDateCoordinates();
+        const i = await this.getPhaseAngle();
+
+        return getApparentMagnitudeMars(coordsHelio.radiusVector, coordsGeo.radiusVector, i);
     }
 }